+++ src/lib/abstraction/init-handler.ts
@@ -1,4 +1,6 @@
-export class InitHandler<T> {
+import {AbstractFormControl} from '../types';
+
+export class InitHandler<T extends AbstractFormControl> {
   private propIsInitialized: boolean = false;
 
   public get isInitialized(): boolean {
@@ -8,15 +10,14 @@ export class InitHandler<T> {
   public setIsInitialized(value: boolean): void {
     const isGroupOrArray = 'controls' in this.instance;
 
-    const ripple = this.propIsInitialized !== value;
     this.propIsInitialized = value;
     if (isGroupOrArray) {
       (this.instance as any as { _setUpControls: () => void })._setUpControls();
     }
-    (this.instance as any).updateValueAndValidity({onlySelf: false, emitEvent: true});
+    this.instance.updateValueAndValidity({onlySelf: false, emitEvent: true});
 
-    if (ripple && isGroupOrArray) {
-      const childControls = (this.instance as any as { controls: { [k: string]: any } }).controls;
+    if (isGroupOrArray) {
+      const childControls = (this.instance as { controls: { [k: string]: any } }).controls;
       Object.keys(childControls)
             .forEach((k) => childControls[k].initHandler.setIsInitialized(value));
     }