+++ src/lib/form-type/abstract-collection-type.ts
@@ -1,5 +1,6 @@
 import {AbstractType} from './abstract-type';
 import {FormArray} from '../abstraction';
+import {ConstructorOf1} from '../types';
 
 export interface ICollectionTypeOptions<T> {
   entrySettings: T;
@@ -10,7 +11,8 @@ export abstract class AbstractCollectionType<TConfig, T> extends AbstractType<T 
 
   public control: FormArray;
 
-  constructor(public readonly childType: new(options: TConfig) => AbstractType<TConfig>,
+  // noinspection TypeScriptAbstractClassConstructorCanBeMadeProtected
+  constructor(public readonly childType: ConstructorOf1<TConfig, AbstractType<TConfig>>,
               options: T & ICollectionTypeOptions<TConfig>) {
     super(options);
     this.fieldInstance = new this.childType(this.options.entrySettings);